pre testing

+++ myapp/controllers/races.py
@@ -5,10 +5,11 @@ from django.http import (
     HttpResponseServerError,
 )
 from django.db import connection
-from ..models import Races, RacesDrivers, SeasonsDrivers, Drivers, Tracks, Teams
+from ..models import Races, RacesDrivers, SeasonsDrivers, Teams
 import json
 from typing import TypedDict, List
 from ..controllers.time import gap_to_time, time_to_seconds
+import traceback
 
 
 class Driver(TypedDict):
@@ -23,6 +24,7 @@ class RaceDriversParams(TypedDict):
 class DriverResult(TypedDict):
     id: str
     time: str
+    plusLaps: int
 
 
 class Results(TypedDict):
@@ -35,32 +37,25 @@ class PostRaceResultsParams(TypedDict):
     results: Results
 
 
-# race name bude treba joinovat s tracks a vytiahnut race_name z tej tabulky, az dropnem name z races
-def getRaces():
-    try:
-        races = Races.objects.values("name", "id")
-
-        data = {"races": []}
-
-        for r in races:
-            data["races"].append({"name": r["name"], "id": str(r["id"])})
-
-        return HttpResponse(json.dumps(data), status=200)
-
-    except Exception as e:
-        print(e)
-
-    return HttpResponseBadRequest()
-
-
 def getRaceDrivers(id: str):
     try:
-        drivers = RacesDrivers.objects.filter(race=id).select_related("driver")
+        drivers = (
+            RacesDrivers.objects.filter(race=id)
+            .select_related("driver", "team")
+            .order_by("team_id")
+        )
 
         data = {"drivers": [], "raceName": ""}
 
         for d in drivers:
-            data["drivers"].append({"id": str(d.driver.id), "name": d.driver.name})
+            data["drivers"].append(
+                {
+                    "id": str(d.driver.id),
+                    "name": d.driver.name,
+                    "teamName": d.team.name,
+                    "logo": d.team.logo,
+                }
+            )
 
         race = Races.objects.select_related("track").get(id=id)
 
@@ -76,20 +71,81 @@ def getRaceDrivers(id: str):
 
 def getEditRace(raceID: str):
     try:
-        race = Races.objects.select_related("track").get(id=raceID)
+        teams = Teams.objects.all().order_by("name")
 
-        teams = Teams.objects.all()
+        race = Races.objects.select_related("track").get(id=raceID)
 
         reservePilot = SeasonsDrivers.objects.filter(
             is_reserve=True, season_id=race.season.id
         ).select_related("driver")
 
-        altTeams = []
+    except Exception as e:
+        print(e)
+        print("no teams found")
+        return HttpResponseServerError()
+
+    # tuna su uz zapisani jazdci na velku cenu
+
+    result = {"raceName": race.track.race_name, "date": str(race.date), "teams": []}
+
+    try:
+        drivers = RacesDrivers.objects.filter(race_id=raceID).select_related(
+            "driver", "team"
+        )
 
         reserve = []
 
+        if len(drivers) > 0:
+            season_drivers = SeasonsDrivers.objects.filter(
+                season_id=race.season_id
+            ).select_related("driver", "team")
+
+            for t in teams:
+                team_drivers = []
+
+                print(t.name)
+
+                for d in drivers:
+                    if d.team.id == t.id:
+                        team_drivers.append(str(d.driver.id))
+                        print(d.driver.name)
+
+                team = {
+                    "teamID": str(t.id),
+                    "teamName": t.name,
+                    "drivers": [],
+                    "color": t.color,
+                    "allOptions": [],
+                }
+
+                for sd in season_drivers:
+                    if sd.is_reserve or sd.team.id == t.id:
+                        team["allOptions"].append(
+                            {
+                                "id": str(sd.driver.id),
+                                "name": sd.driver.name,
+                                "isReserve": sd.is_reserve,
+                            }
+                        )
+
+                for d in team["allOptions"]:
+                    if d["id"] in team_drivers:
+                        team["drivers"].append(d)
+
+                result["teams"].append(team)
+
+            return HttpResponse(json.dumps(result), status=200)
+
+    except Exception as e:
+        traceback.print_exc()
+
+    # ak neni su ziadni zapisani, fetchnem timove dvojice pre sezonu
+
+    try:
         for rp in reservePilot:
-            reserve.append({"id": str(rp.driver.id), "name": rp.driver.name})
+            reserve.append(
+                {"id": str(rp.driver.id), "name": rp.driver.name, "isReserve": True}
+            )
 
         for t in teams:
             drivers = SeasonsDrivers.objects.filter(
@@ -101,19 +157,20 @@ def getEditRace(raceID: str):
                 "teamName": t.name,
                 "drivers": [],
                 "color": t.color,
+                "allOptions": [],
             }
 
             for d in drivers:
-                team["drivers"].append({"id": str(d.driver.id), "name": d.driver.name})
+                team["drivers"].append(
+                    {"id": str(d.driver.id), "name": d.driver.name, "isReserve": False}
+                )
+                team["allOptions"].append(
+                    {"id": str(d.driver.id), "name": d.driver.name, "isReserve": False}
+                )
 
-            altTeams.append(team)
+            team["allOptions"].extend(reserve)
 
-        result = {
-            "raceName": race.track.race_name,
-            "date": str(race.date),
-            "teams": altTeams,
-            "reserves": reserve,
-        }
+            result["teams"].append(team)
 
         return HttpResponse(json.dumps(result), status=200)
 
@@ -123,6 +180,11 @@ def getEditRace(raceID: str):
 
 
 def postEditRaceDrivers(raceID: str, params: RaceDriversParams):
+    if not areDriversUnique(params):
+        return HttpResponseBadRequest(
+            json.dumps({"error": "Jeden pilot nemôže jazdiť za dva tímy naraz"})
+        )
+
     try:
         raceDrivers = RacesDrivers.objects.filter(race_id=raceID)
         raceDrivers.delete()
@@ -202,13 +264,15 @@ def postEditRaceResults(raceID: str, params: PostRaceResultsParams):
 
     data = []
     for d in params["results"]["otherDrivers"]:
-        data.append([gap_to_time(leader["time"], d["time"]), raceID, d["id"]])
+        data.append(
+            [gap_to_time(leader["time"], d["time"]), d["plusLaps"], raceID, d["id"]]
+        )
     try:
         with connection.cursor() as c:
             c.executemany(
                 """
                 UPDATE races_drivers
-                SET time = %s
+                SET time = %s, plus_laps = %s
                 WHERE race_id = %s AND driver_id = %s
                 """,
                 data,
@@ -243,7 +307,7 @@ def getRaceResults(raceID: str):
 					ORDER BY driver_id
                 ),
                 result_times AS (
-                    SELECT r.id AS race_id, d.id, d.name, rd.time + COALESCE(tp.time, 0) AS result_time, has_fastest_lap, t.name AS team_name, r.is_sprint
+                    SELECT r.id AS race_id, d.id, d.name, rd.time + COALESCE(tp.time, 0) AS result_time, has_fastest_lap, t.name AS team_name, r.is_sprint, plus_laps
                     FROM races AS r
                     JOIN races_drivers AS rd ON r.id = rd.race_id
                     LEFT JOIN time_penalties AS tp ON tp.driver_id = rd.driver_id
@@ -253,14 +317,14 @@ def getRaceResults(raceID: str):
                     ORDER BY result_time
                 )
 
-                SELECT id, name, result_time, ROW_NUMBER() OVER (PARTITION BY rt.race_id ORDER BY result_time ASC), has_fastest_lap, team_name, is_sprint
+                SELECT id, name, result_time, ROW_NUMBER() OVER (PARTITION BY rt.race_id ORDER BY result_time ASC), has_fastest_lap, team_name, is_sprint, plus_laps
                 FROM result_times AS rt
                 ORDER BY result_time ASC
             """,
             [raceID, raceID],
         )
 
-        # [0: driver_id, 1: driver_name, 2: result_time, 3: rank, 4: has_fastest_lap, 5: team_name, 6: is_sprint]
+        # [0: driver_id, 1: driver_name, 2: result_time, 3: rank, 4: has_fastest_lap, 5: team_name, 6: is_sprint, 7: plus_laps]
         drivers = c.fetchall()
 
         results = {"results": []}
@@ -275,6 +339,7 @@ def getRaceResults(raceID: str):
                     "rank": d[3],
                     "hasFastestLap": d[4],
                     "isSprint": d[6],
+                    "plusLaps": d[7],
                 }
             )
 
@@ -283,3 +348,15 @@ def getRaceResults(raceID: str):
     except Exception as e:
         print(e)
         return HttpResponseBadRequest()
+
+
+def areDriversUnique(params: RaceDriversParams) -> bool:
+    all_drivers = []
+
+    for team in params["teams"]:
+        for driver in team["drivers"]:
+            if driver in all_drivers:
+                return False
+            all_drivers.append(driver)
+
+    return True
